Got reading of shaders working.
This required a lot of work and several new packages. We added
WebPack and a host of plugins for it so that we could use the ES6
require syntax to include the contents of our shaders.

This brings the total of the applications we use to perform our
build up to three: one to just build the web app bundle, another
to build our server script, and npm to manage those processes.

God I hate the JavaScript ecosystem.